Fix comparer helper collisions in ContainsAll PR
Reuse the existing non-generic comparer adapter and shared case-insensitive test comparer after main introduced equivalent helpers, which resolves the PR's duplicate type CI failures.\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
